pyproject.toml
[tool.poetry]
name = "salesight"
version = "0.1.0"
description = "ai project"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.7.4"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
colorama = "0.4.6"
crashtest = "0.4.1"
distlib = "0.3.8"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.15.4"
idna = "3.7"
importlib-metadata = "8.2.0"
installer = "0.7.0"
jaraco-classes = "3.4.0"
keyring = "24.3.1"
more-itertools = "10.3.0"
msgpack = "1.0.8"
packaging = "24.1"
pexpect = "4.9.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
ptyprocess = "0.7.0"
pyproject-hooks = "1.1.0"
pywin32-ctypes = "0.2.2"
rapidfuzz = "3.9.5"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
tomlkit = "0.13.0"
trove-classifiers = "2024.7.2"
urllib3 = "2.2.2"
virtualenv = "20.26.3"
zipp = "3.19.2"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cleanup = "^8.1.0"
pandas = "^2.2.2"
supabase = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
